feat: enhance PostgreSQL search functionality with new SQL helpers and scoring methods
- Added `EnsureSearchSqlHelpers` method to create necessary SQL functions for improved search capabilities.
- Introduced new database functions for word similarity and case-insensitive LIKE operations, enhancing fuzzy search accuracy.
- Updated `PostgresFuzzySearchProvider` to utilize the new search functions, improving query performance and results.
- Refactored `FranchiseTitleHelper` and `MovieSimilarityWeights` to accommodate changes in similarity scoring logic.
- Enhanced unit and integration tests to validate the new search functionalities and ensure robustness.

+++ Jellyfin.Plugin.Pgsql/Similar/FranchiseTitleHelper.cs
@@ -99,7 +99,9 @@ public static partial class FranchiseTitleHelper
     /// <returns>Franchise score contribution, or 0 when below the floor.</returns>
     public static int FranchiseScoreFromWordSimilarity(double wordSimilarity)
     {
-        if (wordSimilarity < MovieSimilarityWeights.TitleWordSimilarityFloor)
+        // Compare as float to match the SQL floor (pg_trgm / EF float) and avoid
+        // double rounding surprises around 0.4 (0.4d < 0.4f when promoted).
+        if ((float)wordSimilarity < MovieSimilarityWeights.TitleWordSimilarityFloor)
         {
             return 0;
         }
